# Performance improvements in LINQ methods All() & Any() for List<T> & Array