@Clockwork-Muse 2020-01-16
```csharp
[Benchmark]
public bool AllNewOnAll()
{
	Func<int, bool> fp = n => n < Size;
	return _range.All(fp) && _list.All(fp) && _array.All(fp);
}

[Benchmark]
public bool AllOriginalOnAll()
{
	Func<int, bool> fp = n => n < Size;
	return System.Linq.Enumerable.All(_range, fp) && System.Linq.Enumerable.All(_list, fp) && System.Linq.Enumerable.All(_array, fp);
}
```

uh, I assume you've defined an `Any()`/`All()` somewhere else, which makes it hard for us to duplicate your results.  I believe the normal pattern is to just define an `Benchmark` over the relevant, real, methods, then running the same thing over both the original and the modified code.  It also helps to prevent certain JIT shenanigans. 

You should probably be splitting the test by the type chosen - as it is, performance changes in one type can be hidden by changes in another.
Additionally, you should probably be parameterizing `Limit`, as well as whether the methods would return `true`/`false`, to help eliminate the effects of pipeline caching.
Without having tried it, I _believe_ using [`IntoArgumentsSource`](https://benchmarkdotnet.org/articles/features/parameterization.html#sample-introargumentssource) (in combination with `Params`) would allow you to eliminate the use of `GlobalSetup`, and just pass in an `IEnumerable<T>`, making your test code otherwise quite a bit cleaner.