@jvandertil 2020-01-21
Just for reference, I have submitted something similar in the past, but we had a hard time finding places where this makes a measurable impact. See: https://github.com/dotnet/corefx/pull/39200