@stephentoub 2020-02-17
Thank you for the PR. However, I'm going to close this. If you're handed a `List<T>` or a `T[]`, as you pointed out the right thing to do is to use the methods on those types that do this more efficiently. While it's certainly possible you're handed one typed as `IEnumerable<T>`, none of the other LINQ methods do so, and C# iterators won't produce such types, so it's not the common case. Adding these special-cases then penalizes the common case by making them pay for the type checks when they won't add benefit, and that can be measurable for the common case of enumerables with just a few elements. At present it's just not the right trade-off. Regardless, thank you for the contribution.